requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.21.0
APScheduler==3.11.1
attrs==25.4.0
cffi==2.0.0
flatbuffers==25.9.23
frozenlist==1.8.0
h3==4.3.1
idna==3.11
multidict==6.7.0
numpy==2.3.5
propcache==0.4.1
pyaes==1.6.1
pyasn1==0.6.1
pycparser==2.23
python-dotenv==1.2.1
python-socks==2.7.3
pytz==2025.2
rsa==4.9.1
Telethon==1.42.0
timezonefinder==8.1.0
typing_extensions==4.15.0
tzlocal==5.3.1
yarl==1.22.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
pygeohash==1.2.0
redis[hiredis]==5.0.1
polyline==2.0.2

